What we offer - build fix - move styles.ts from ourAreas to the components directory
