+++ src/components/whatWeDo/ourDevelopmentAreas/OurDevelopmentAreas.tsx
@@ -1,13 +1,13 @@
 import React from "react"
 import styled from "styled-components"
-import Blockchain from "../../assets/blockchain.svg"
-import MobileAppDevelopment from "../../assets/mobileAppDevelopment.svg"
-import ProductDesign from "../../assets/productDesign.svg"
-import RightArrow from "../../assets/rightArrow.svg"
-import WebDevelopment from "../../assets/webDevelopment.svg"
-import AgileWorkshops from "../../assets/agileWorkshops.svg"
-import CustomSoftwareDevelopment from "../../assets/customSoftwareDevelopment.svg"
-import variables from "../../styles/variables"
+import Blockchain from "../../../assets/blockchain.svg"
+import MobileAppDevelopment from "../../../assets/mobileAppDevelopment.svg"
+import ProductDesign from "../../../assets/productDesign.svg"
+import RightArrow from "../../../assets/rightArrow.svg"
+import WebDevelopment from "../../../assets/webDevelopment.svg"
+import AgileWorkshops from "../../../assets/agileWorkshops.svg"
+import CustomSoftwareDevelopment from "../../../assets/customSoftwareDevelopment.svg"
+import variables from "../../../styles/variables"
 import { Link } from "gatsby"
 
 const DevelopmentAreasWrapper = styled.div({
