+++ src/components/whatWeDo/ourDevelopmentAreas/styles.ts
@@ -1,5 +1,5 @@
 import styled from "styled-components"
-import variables from "../../styles/variables"
+import variables from "../../../styles/variables"
 
 export const Container = styled.div({
   marginTop: "90px",
