+++ src/pages/our-areas/web-development.tsx
@@ -10,7 +10,7 @@ import {
   DescriptionWrapper,
   Paragraph,
   Title,
-} from "./styles"
+} from "../../components/whatWeDo/ourDevelopmentAreas/styles"
 
 const WebDevelopment = () => {
   return (
