+++ src/pages/what-we-offer.tsx
@@ -5,7 +5,7 @@ import Ratings from "../components/subcomponents/Ratings"
 import Banners from "../components/whatWeDo/banners/Banners"
 import Contact from "../components/whatWeDo/contact/Contact"
 import HowWeWork from "../components/whatWeDo/howWeWork/HowWeWork"
-import OurDevelopmentAreas from "../components/whatWeDo/OurDevelopmentAreas"
+import OurDevelopmentAreas from "../components/whatWeDo/ourDevelopmentAreas/OurDevelopmentAreas"
 import Showcase from "../components/whatWeDo/showcase/Showcase"
 import { Page } from "../layout/Page"
 import variables from "../styles/variables"
